+++ boards/opentitan/src/io.rs
@@ -4,9 +4,7 @@ use core::str;
 use kernel::debug;
 use kernel::debug::IoWrite;
 
-use crate::CHIP;
-use crate::PROCESSES;
-use crate::PROCESS_PRINTER;
+use crate::setup::{CHIP, PROCESSES, PROCESS_PRINTER};
 
 struct Writer {}
