+++ boards/opentitan/src/tests/mod.rs
@@ -1,8 +1,4 @@
-use crate::BOARD;
-use crate::CHIP;
-use crate::MAIN_CAP;
-use crate::NUM_PROCS;
-use crate::PLATFORM;
+use crate::setup::{BOARD, CHIP, MAIN_CAP, NUM_PROCS, PLATFORM};
 use kernel::debug;
 
 pub fn semihost_command_exit_success() -> ! {
