+++ boards/opentitan/src/tests/otbn.rs
@@ -1,5 +1,5 @@
+use crate::setup::PERIPHERALS;
 use crate::tests::run_kernel_op;
-use crate::PERIPHERALS;
 use core::cell::Cell;
 use kernel::static_init;
 use kernel::utilities::cells::TakeCell;
